bot-desafio-tecnico-havan/requirements.txt
botcity-framework-web>=1.0,<2.0
botcity-maestro-sdk>=0.5.1,<1.0
botcity-excel-plugin
webdriver-manager
# Uncomment to use botcity-utils
# botcity-utils>=0.1.2

# Uncomment to use botcity-crawler-plugin
# botcity-crawler-plugin

# Uncomment to use Pandas
pandas
